Added abstract for media file searcher
+ moved getmediafiles to abstract
+ will be used for other things to
